# cmd/api/docs/swagger.yaml
definitions:
...
  response.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: boolean
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server  server. You can visit the GitHub repository
    at https://github.com/prasanth-pn/clean-code-architecture-ecommerce
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Go + Gin ecommerce API
  version: "1.0"
paths:
...
  /admin/add/products:
    post:
      operationId: AdminAddProducts
      parameters:
      - description: select the  image
        in: formData
        name: homepic
        required: true
        type: file
      - description: AdminAddProduct
        in: formData
        name: products
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: AddProducts for admin
      tags:
      - PRODUCTMANAGEMENT
  /admin/generate-coupon:
    post:
      operationId: GenereateCoupon for admin
      parameters:
      - description: coupon
        in: query
        name: coupon
        required: true
        type: string
      - description: quantity
        in: query
        name: quantity
        required: true
        type: string
      - description: validity
        in: query
        name: validity
        required: true
        type: string
      - description: discount
        in: query
        name: discount
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: GenerateCoupon for admin
      tags:
      - COUPON
...
  /admin/list/category:
    get:
      operationId: ListCategories for admin
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: string
      - description: pagesize
        in: query
        name: pagesize
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: ListCategories for admin
      tags:
      - PRODUCTMANAGEMENT
  /admin/list/productby-categories:
    get:
      operationId: listproductsbycategories for admin
      parameters:
      - description: category_id
        in: query
        name: cat_id
        required: true
        type: string
      - description: page
        in: query
        name: page
        required: true
        type: string
      - description: pagesize
        in: query
        name: pagesize
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: ListProductByCategories for admin
      tags:
      - PRODUCTMANGEMENT
...
  /admin/product/delete:
    delete:
      operationId: AdminAdddProducts for admin
      parameters:
      - description: product_id
        in: query
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: AdminDeleteProducts for admin
      tags:
      - PRODUCTMANAGEMENT
  /admin/products/delete-image:
    delete:
      operationId: DeleteImageFromProduct for admin
      parameters:
      - description: select the image name
        in: query
        name: imageName
        required: true
        type: string
      - description: enter the prouduct Id
        in: query
        name: product_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: DeleteImagein Products
      tags:
      - PRODUCTMANAGEMENT
...
  /admin/update/product:
    patch:
      operationId: UpdateProduct for admin
      parameters:
      - description: product_id
        in: query
        name: product_id
        required: true
        type: string
      - description: select image
        in: formData
        name: image
        required: true
        type: file
      - description: updateproduct
        in: formData
        name: updateproduct
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: UpdateProduct for admin
      tags:
      - PRODUCTMANAGEMENT
  /admin/upload/image:
    patch:
      operationId: productImageUpload for admin
      parameters:
      - description: product_id
        in: query
        name: product_id
        required: true
        type: string
      - description: select image
        in: formData
        name: image
        required: true
        type: file
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: ImageUploadproduct for admin
      tags:
      - PRODUCTMANAGEMENT
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"